+++ src/viur_cli/__init__.py
@@ -1,10 +1,10 @@
 from .version import *
 
 from .cli import *
+from .npm import *
+from .release import *
 from .flare import *
 from .deploy import *
 from .local import *
 from .vi import *
-from .assets import *
-from .vue import *
 from .setup import *